examples/open-deep-research/requirements.txt
galadriel>=0.0.10
beautifulsoup4==4.12.3
Bio
chess
datasets==2.21.0
google-search-results==2.4.2
mammoth==1.8.0
markdownify==0.14.1
numexpr==2.10.1
numpy>=1.26.4,<2.0.0
openai>=1.52.2
openpyxl==3.1.2
pandas==2.2.3
pathvalidate==3.2.1
pdfminer==20191125
pdfminer.six==20240706
Pillow==11.0.0
pubchempy==1.0.4
puremagic==1.28
PyPDF2==3.0.1
pypdf==5.1.0
python-dotenv==1.0.1
python-pptx==1.0.2
pydub==0.25.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
serpapi==0.1.5
SpeechRecognition==3.10.1
sympy==1.12
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
transformers==4.46.0
xlrd==2.0.1
youtube-transcript-api==0.6.2